cast length specifier.
2005-01-21  Not Zed  <NotZed@Ximian.com>

        * util/eab-book-util.c (eab_name_and_email_query): cast length
        specifier.

svn path=/branches/gnome-2-8/; revision=28473

+++ addressbook/util/eab-book-util.c
@@ -108,7 +108,7 @@ eab_name_and_email_query (EBook *book,
 		while (*t && *t != '@')
 			++t;
 		if (*t == '@' && escaped_name) {
-			email_query = g_strdup_printf ("(beginswith \"email\" \"%.*s@\")", t-escaped_email, escaped_email);
+			email_query = g_strdup_printf ("(beginswith \"email\" \"%.*s@\")", (int)(t-escaped_email), escaped_email);
 
 		} else {
 			email_query = g_strdup_printf ("(beginswith \"email\" \"%s\")", escaped_email);
